// src/therun/types.ts
export type RaceStatus =
  | "pending"
  | "starting"
  | "progress"
  | "finished"
  | "aborted";
export type RaceParticipantStatus =
  | "joined"
  | "ready"
  | "started"
  | "finished"
  | "confirmed"
  | "abandoned"
  | "unjoined";

export type RaceStartMethodType = "everyone-ready" | "datetime" | "moderator";

export interface Race {
  raceId: string;
  creator: string;
  createdAt: string;
  game: string;
  displayGame: string;
  category: string;
  displayCategory: string;
  gameImage: string;
  description: string;
  canStartEarly: boolean;
  status: RaceStatus;
  customRules: RaceRule[];
  customName: string;
  visible: boolean;
  // Not available in paginated races due to too much data
  participants?: RaceParticipantWithLiveData[];
  // Only available in paginated races because pending races have no results yet
  results: RaceResult[] | null;
  moderators: string[];
  participantCount: number;
  readyParticipantCount: number;
  finishedParticipantCount: number;
  prizepool?: number;
  prizepoolDistribution?: number[];
  startTime: string | null;
  endTime: string | null;
  firstFinishedParticipantTime: number | null;
  isTestRace: boolean;
  increment?: number;
  gameIncrement?: number;
  isFeatured: boolean;
  nextRaceId?: string;
  previousRaceId?: string;
  forceStream?: string;
  requiresPassword?: boolean;
  ranked: boolean;
  autoConfirm: boolean;
  countdownSeconds: number;
  startMethod: RaceStartMethodType;
  willStartAt?: string | null;
  timeLeaderboards: RaceTimeStat[];
  mmrLeaderboards: RaceMmrStat[];
}

export interface RaceResult {
  position: number;
  name: string;
  status: RaceParticipantStatus;
  finalTime: number | null;
}

export type RaceParticipant = {
  raceId: string;
  user: string;
  status: RaceParticipantStatus;
  pb: string;
  disqualified: boolean;
  disqualifiedBy: string | null;
  disqualifiedReason: string | null;
  finalTime: number | null;
  joinedAtDate: string | null;
  readyAtDate: string | null;
  finishedAtDate: string | null;
  confirmedAtDate: string | null;
  abandondedAtDate: string | null;
  ratingBefore: number;
  ratingAfter: number | null;
  comment?: string | null;
};

export interface RaceParticipantWithLiveData extends RaceParticipant {
  liveData?: RaceLiveData;
}

export interface RaceLiveData {
  // How far along is a runner if you look at the split amount (useful when theres no pb yet, to at least have an indication)
  runPercentageSplits: number;
  // How far along is a runner compared to pb (best way to see how far someone is). Can be null if runner has no pb yet
  runPercentageTime: number | null;
  currentSplitIndex: number;
  totalSplits: number;
  currentTime: number;
  gameTime: boolean;
  // Important: with this info we can know how far into a split the runner is. Useful to infer percentage in between splits.
  startedAt: number;
  splitStartedAt: number;
  currentSplitName: string;
  // Not very stable, but can use it
  currentPrediction: number | null;
  streaming: boolean;
  // +- to pb currently
  delta: number;
  runFinished: boolean;
  bestPossibleTime: number | null;
  timeToNextSplit?: number;
}

// TODO:: determine which rules are possible
export interface RaceRule {}
// ...
export interface RaceTimeStat {
  user: string;
  time: number;
}

export interface RaceMmrStat {
  user: string;
  mmr: number;
}

export interface RaceRanking<T extends "mmr" | "time" | "timeMonth"> {
  score: number;
  rank: number;
  type: T;
}

export interface User {
  bio: string;
  user: string;
  picture: string;
  login: string;
  prounouns: string;
  searchName: string;
}